+++ config/routes.rb
@@ -30,10 +30,10 @@ Rails.application.routes.draw do
   end
 
   scope module: :analytics, path: :analytics do
-    match '' => redirect('analytics/general'), via: [:get]
-    match 'general' => 'general#index', via: [:get], as: :general_analytics
-    match 'subject' => 'subject#index', via: [:get], as: :subject_analytics
-    match 'group' => 'group#index', via: [:get], as: :group_analytics
+    get '' => redirect('analytics/general')
+    get 'general' => 'general#index', as: :general_analytics
+    get 'subject' => 'subject#index', as: :subject_analytics
+    get 'group' => 'group#index', as: :group_analytics
   end
 
   resources :users, only: %i[index new create destroy show] do